Move job-proxy.service to proxy folder
