+++ src/app/components/job-state-status-bar/job-state-status-bar.component.ts
@@ -4,7 +4,7 @@ import {Job} from "../../models/job/job";
 import {UserManager} from "../../services/user-manager.service";
 import {UserProxy} from "../../services/proxy/user-proxy.service";
 import {UserJob} from "../../models/user/user-job";
-import {JobProxy} from "../../services/job-proxy.service";
+import {JobProxy} from "../../services/proxy/job-proxy.service";
 import {Router} from "@angular/router";
 import {UserBankAccount} from "../../models/user/user-bank-account";
