+++ src/app/views/job-details/job-details.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from "@angular/core";
 import {ActivatedRoute} from "@angular/router";
 import {Job} from "../../models/job/job";
-import {JobProxy} from "../../services/job-proxy.service";
+import {JobProxy} from "../../services/proxy/job-proxy.service";
 import {UserManager} from "../../services/user-manager.service";
 import {User} from "../../models/user";
