+++ src/app/views/jobs/jobs.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from "@angular/core";
 import {Job} from "../../models/job/job";
-import {JobProxy} from "../../services/job-proxy.service";
+import {JobProxy} from "../../services/proxy/job-proxy.service";
 import {Location} from "@angular/common";
 import {ActivatedRoute} from "@angular/router";
