+++ src/app/views/my-jobs/company-jobs/company-jobs.component.ts
@@ -1,6 +1,6 @@
 import {Component, Input, OnInit} from "@angular/core";
 import {Job} from "../../../models/job/job";
-import {JobProxy} from "../../../services/job-proxy.service";
+import {JobProxy} from "../../../services/proxy/job-proxy.service";
 import {UserManager} from "../../../services/user-manager.service";
 
 @Component({
